+++ internal/migrator/migrations/000001_init.up.sql
@@ -12,7 +12,7 @@ CREATE TABLE
         user_id INT,
         number TEXT,
         process TEXT,
-        accrual FLOAT (2),
+        accrual FLOAT (5),
         uploaded_at TEXT,
         FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
     );
@@ -21,6 +21,6 @@ CREATE TABLE
     IF NOT EXISTS balance (
         id SERIAL PRIMARY KEY,
         user_id INT,
-        balance FLOAT (2),
+        balance FLOAT (5),
         FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
     );
