+++ internal/usecase/update_balance.go
@@ -1,6 +1,6 @@
 package usecase
 
-func (usecase *Usecase) UpdateBalance(balance float32, userID string) error {
+func (usecase *Usecase) UpdateBalance(balance float64, userID string) error {
 	err := usecase.store.UpdateBalance(balance, userID)
 
 	return err
